Task 5/requirements.txt
langchain
langchain-groq
chromadb
pypdf
wikipedia
python-dotenv
